04 05 06 07 validated

+++ 09/ex02/PmergeMe.hpp
@@ -6,7 +6,7 @@
 /*   By: fle-tolg <fle-tolg@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/30 13:06:28 by fle-tolg          #+#    #+#             */
-/*   Updated: 2023/05/01 14:58:35 by fle-tolg         ###   ########.fr       */
+/*   Updated: 2023/05/06 10:55:05 by fle-tolg         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,6 +14,7 @@
 # define PMERGEME_HPP
 
 # include <iostream>
+# include <ostream>
 # include <list>
 # include <vector>
 # include <algorithm>
@@ -21,9 +22,9 @@
 class PmergeMe
 {
 	private:
-		std::vector<long int>	_arg;
-		std::vector<long int>	_vector;
-		std::list<long int>		_list;
+		std::vector<int>	_arg;
+		std::vector<int>	_vector;
+		std::list<int>		_list;
 		unsigned long int 		_timeVector;
 		unsigned long int 		_timeList;
